requirements.txt
channels==4.0.0
Django==4.2
djangorestframework==3.14.0
daphne==4.0.0

